+++ src/core/lib.zig
@@ -4,6 +4,7 @@ pub const hash = @import("hash/hash.zig");
 pub const merkle = @import("merkle/merkle.zig");
 pub const channel = @import("channel/channel.zig");
 pub const serialization = @import("serialization.zig");
+pub const pool = @import("pool.zig");
 pub const bit_utils = @import("bit_utils.zig");
 pub const simd = @import("simd.zig");